# schemas/microduck-simulator-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: MicroDuck ROM Simulator API
  version: 1.0.0
paths:
...
  /v1/catalog:
    get:
      summary: Catalog
      operationId: catalog
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
  /v1/robot/status:
    get:
      summary: Robot Status
      operationId: robotStatus
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RobotStatus'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
  /v1/tasks:
    post:
      summary: Create Task
      operationId: createTask
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSnapshot'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request Entity Too Large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '503':
          description: Service Unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
      - bearerAuth: []
  /v1/tasks/{taskId}:
    get:
      summary: Get Task
      operationId: getTask
      security:
      - bearerAuth: []
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          pattern: ^[0-9a-f]{32}$
          title: Taskid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSnapshot'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Service Unavailable
  /v1/tasks/{taskId}/cancel:
    post:
      summary: Cancel Task
      operationId: cancelTask
      security:
      - bearerAuth: []
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          pattern: ^[0-9a-f]{32}$
          title: Taskid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSnapshot'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Service Unavailable
  /v1/tasks/{taskId}/command:
    put:
      summary: Command Task
      operationId: commandTask
      security:
      - bearerAuth: []
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          pattern: ^[0-9a-f]{32}$
          title: Taskid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCommandRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskSnapshot'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Conflict
        '413':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Request Entity Too Large
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Service Unavailable
  /v1/tasks/{taskId}/events:
    get:
      summary: Task Events
      operationId: taskEvents
      security:
      - bearerAuth: []
      parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          pattern: ^[0-9a-f]{32}$
          title: Taskid
      - name: afterSequence
        in: query
        required: false
        schema:
          type: integer
          minimum: -1
          default: -1
          title: Aftersequence
      - name: pageSize
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Pagesize
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskEventPage'
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Not Found
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Service Unavailable
components:
  schemas:
...
    CatalogResponse:
      properties:
        bundleId:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Bundleid
        bundleVersion:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Bundleversion
        bundleDigest:
          anyOf:
          - type: string
            pattern: ^sha256:[0-9a-f]{64}$
          - type: 'null'
          title: Bundledigest
        observationContract:
          type: string
          const: MICRODUCK_OBS_61_V1
          title: Observationcontract
          default: MICRODUCK_OBS_61_V1
        actionContract:
          type: string
          const: MICRODUCK_ACTION_14_V1
          title: Actioncontract
          default: MICRODUCK_ACTION_14_V1
        actions:
          items:
            $ref: '#/components/schemas/ActionDefinition'
          type: array
          maxItems: 256
          title: Actions
      additionalProperties: false
      type: object
      required:
      - actions
      title: CatalogResponse
...
    Error:
      properties:
        code:
          type: string
          enum:
          - AUTH_REQUIRED
          - NOT_READY
          - BUNDLE_MISMATCH
          - ACTION_UNAVAILABLE
          - PARAMETER_INVALID
          - PRECONDITION_FAILED
          - ROBOT_BUSY
          - TASK_ID_CONFLICT
          - COMMAND_SEQUENCE_CONFLICT
          - STALE_COMMAND
          - TASK_NOT_FOUND
          - REQUEST_BODY_TOO_LARGE
          - INTERNAL_ERROR
          title: Code
        message:
          type: string
          maxLength: 256
          minLength: 1
          title: Message
        details:
          additionalProperties:
            $ref: '#/components/schemas/JsonValue'
          propertyNames:
            maxLength: 128
            minLength: 1
          type: object
          maxProperties: 32
          title: Details
      additionalProperties: false
      type: object
      required:
      - code
      - message
      - details
      title: Error
...
    RobotStatus:
      properties:
        schema:
          type: string
          const: BIPED_POSE_V1
          title: Schema
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        basePositionM:
          prefixItems:
          - type: number
          - type: number
          - type: number
          type: array
          maxItems: 3
          minItems: 3
          title: Basepositionm
        baseOrientationXyzw:
          prefixItems:
          - type: number
          - type: number
          - type: number
          - type: number
          type: array
          maxItems: 4
          minItems: 4
          title: Baseorientationxyzw
        baseLinearVelocityMps:
          prefixItems:
          - type: number
          - type: number
          - type: number
          type: array
          maxItems: 3
          minItems: 3
          title: Baselinearvelocitymps
          description: World-frame trunk-base linear velocity.
        baseAngularVelocityRadps:
          prefixItems:
          - type: number
          - type: number
          - type: number
          type: array
          maxItems: 3
          minItems: 3
          title: Baseangularvelocityradps
          description: Trunk-body-frame angular velocity, matching the training IMU
            gyro.
        jointPositionsRad:
          prefixItems:
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          type: array
          maxItems: 14
          minItems: 14
          title: Jointpositionsrad
        jointVelocitiesRadps:
          prefixItems:
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          - type: number
          type: array
          maxItems: 14
          minItems: 14
          title: Jointvelocitiesradps
        policyTarget:
          additionalProperties:
            $ref: '#/components/schemas/JsonValue'
          propertyNames:
            maxLength: 128
            minLength: 1
          type: object
          maxProperties: 32
          title: Policytarget
        requestedMotion:
          additionalProperties:
            $ref: '#/components/schemas/JsonValue'
          propertyNames:
            maxLength: 128
            minLength: 1
          type: object
          maxProperties: 32
          title: Requestedmotion
        appliedMotion:
          additionalProperties:
            $ref: '#/components/schemas/JsonValue'
          propertyNames:
            maxLength: 128
            minLength: 1
          type: object
          maxProperties: 32
          title: Appliedmotion
        limitingReason:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Limitingreason
        activePolicyRef:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Activepolicyref
        activeActionCode:
          anyOf:
          - type: string
            maxLength: 128
            minLength: 1
          - type: 'null'
          title: Activeactioncode
        activeTaskId:
          anyOf:
          - type: string
            pattern: ^[0-9a-f]{32}$
          - type: 'null'
          title: Activetaskid
        simulationTimeS:
          type: number
          minimum: 0.0
          title: Simulationtimes
        loopFrequencyHz:
          type: number
          minimum: 0.0
          title: Loopfrequencyhz
        fallen:
          type: boolean
          title: Fallen
        limp:
          type: boolean
          title: Limp
        health:
          additionalProperties:
            $ref: '#/components/schemas/JsonValue'
          propertyNames:
            maxLength: 128
            minLength: 1
          type: object
          maxProperties: 32
          title: Health
      additionalProperties: false
      type: object
      required:
      - schema
      - timestamp
      - basePositionM
      - baseOrientationXyzw
      - baseLinearVelocityMps
      - baseAngularVelocityRadps
      - jointPositionsRad
      - jointVelocitiesRadps
      - policyTarget
      - requestedMotion
      - appliedMotion
      - simulationTimeS
      - loopFrequencyHz
      - fallen
      - limp
      - health
      title: RobotStatus
...
    TaskCreateRequest:
      properties:
        schema:
          type: string
          const: MICRODUCK_SIM_TASK_V1
          title: Schema
        taskId:
          type: string
          pattern: ^[0-9a-f]{32}$
          title: Taskid
        actionCode:
          type: string
          maxLength: 128
          minLength: 1
          title: Actioncode
        bundleVersion:
          type: string
          maxLength: 128
          minLength: 1
          title: Bundleversion
        bundleDigest:
          type: string
          pattern: ^sha256:[0-9a-f]{64}$
          title: Bundledigest
        parameters:
          additionalProperties:
            anyOf:
            - type: string
              maxLength: 1024
            - type: integer
            - type: number
            - type: boolean
            - type: 'null'
          propertyNames:
            maxLength: 64
            minLength: 1
          type: object
          maxProperties: 16
          title: Parameters
        scenario:
          $ref: '#/components/schemas/Scenario'
        leaseMs:
          anyOf:
          - type: integer
            maximum: 60000.0
            exclusiveMinimum: 0.0
          - type: 'null'
          title: Leasems
        requestedBy:
          type: string
          maxLength: 256
          minLength: 1
          title: Requestedby
      additionalProperties: false
      type: object
      required:
      - schema
      - taskId
      - actionCode
      - bundleVersion
      - bundleDigest
      - parameters
      - scenario
      - requestedBy
      title: TaskCreateRequest
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-token